Split graphs into serial and parallel versions
We were previously using the same graph for both parallel and serial
algorithms. This was kind of ok when there was only one atomic variable
to access, but now that there must be more it makes sense to split
the graph types into a serial and a paralllel version.

This also makes it easier to experiment with other graph representations
other than CSR, such as Linear.