WIP: Add serial SP2, parallel SP1 and two versions of parallel SP2 (edges and vertices)
After quite a bit of experimentation I finally was able to get
a semi-competitive version of the algorithms implemented. Serial
SP2 is just added for comparison.

The parallel version of the algorithms uses a single for_each loop
where it does everything, including popping ang pushing to the heap.

Depending on the graph this version demonstrates a modest speedup
as more threads are added and can be faster than the serial version.

WIP: as this requires actual run results so we can compare.