Add very lightweight profiling to SSSP algorithms
This adds much more lightweight time profiling to the
sssp algorithms and prints the results in the end.

In my server, on the kronecker graphs this showed two
things.
1 - As we suspected the median results for SP1 and SP2
(and even dijsktra) are much better than for delta step
where the nodes are only really fixed when the computation
completes (for instance for kron 2e24, this means we're
about 50% faster than delta step, and 20-30% faster than
dijskstra).
2 - The time that it takes to fix the last node is actually
about half the total runtime, which means:
a) we're wasting a lot of work somewhere, like taking
duplicates off the heap and...
b) perhaps more importantly, we've been measuring the
wrong thing. What we care about is the time that it
takes to fix all the nodes that can be fixed, not some
arbitrary total runtime, so we should use that max
time to fixedness as the new total runtime.